Move file-settings and keystore reconciliation back to the stateful driver
Stateless clusters don't use the init container keystore (they deliver
secure settings via cluster_secrets in file-based settings) and manage
the file-settings Secret themselves via filesettings.ReconcileClusterSecrets.
Running either in ReconcileSharedResources is useless work for stateless
and, for file-settings, conflicts with SCP-deferral logic.

Restore the original per-driver approach: the stateful driver explicitly
calls MaybeReconcileEmptyFileSettingsSecret (version-gated) and its own
reconcileKeystore (with managed keystore password and remote API keys).
Both are stateful-only now. The stateless driver is untouched.
